+++ test/features/multiple.js
@@ -132,6 +132,25 @@ define(function () {
         expect($('#element').parsley().__class__).to.be('ParsleyFieldMultiple');
         expect($('#element').attr('data-parsley-multiple')).to.be('element');
       });
+      it('should remove errors on change, whatever field is changed', function () {
+        $('body').append(
+          '<form id="element" >' +
+            '<input type="checkbox" name="check[]" id="check1" value="1" required data-parsley-mincheck="1" />'  +
+            '<input type="checkbox" name="check[]" id="check2" value="2" />'  +
+            '<input type="checkbox" name="check[]" id="check3" value="3" />'  +
+            '<input type="checkbox" name="check[]" id="check4" value="4" />'  +
+          '</form>');
+        var parsleyInstance = $('#check1').parsley();
+        parsleyInstance.validate();
+        expect(parsleyInstance.$elements.length).to.be(4);
+        expect(parsleyInstance.validationResult).not.to.be(true);
+        $('#check2').attr('checked', 'checked');
+        $('#check2').trigger($.Event('change'));
+        expect(parsleyInstance.validationResult).to.be.eql(true);
+      });
+      it.skip('should add errors on change if trigger enabled, whatever field is changed', function () {
+
+      });
       afterEach(function () {
         window.ParsleyConfig = { i18n: window.ParsleyConfig.i18n, validators: window.ParsleyConfig.validators };